Still bug fixing... crashes on initial calc
Really frustratingly, I can't seem to find where the code is excuting,
or there is something odd and it is crashng so hard it never gets to
cout.
it could also be that it is crashing while reading fort.7 or similar,
and this is happening at class initialisation so you never actually get
to the function where the calculate J code is?